✨ Modularize OpenRPC Viewer with new components and hooks
- Extract FileTreeNode, MethodDetail, and DropdownMenu into separate components
- Create useFileTree hook to manage file tree state and logic
- Simplify OpenRPCViewer component by delegating responsibilities
- Improve code organization and reusability
